[refactor] : Home component moved to features/Home from pages
