+++ src/features/Home/index.js
@@ -3,3 +3,4 @@ export { default as FitnessScore } from './FitnessScore';
 export { default as LeaderBoard } from './LeaderBoard';
 export { default as DuePaymentIndicator } from './DuePaymentIndicator';
 export { default as MoveCoins } from './MoveCoins';
+export { default as Home } from './Home';
