+++ src/pages/index.js
@@ -1,5 +1,4 @@
 export { default as Login } from './Login';
 export { default as LoadingPage } from './LoadingPage';
-export { default as Home } from './Home';
 export { default as PageNotFound } from './PageNotFound';
 export { default as OpeningPage } from './OpeningPage';
